Set test.conf to use a useless in-memory postgres database so by default WithApplication has a useless database, and revert WithPostgresContainer for tests that require a real postgres database.
