+++ universal-application-tool-0.0.1/conf/application.test.conf
@@ -1,12 +1,7 @@
 include "application.conf"
 
 db {
-  default.driver = "org.testcontainers.jdbc.ContainerDatabaseDriver"
-  # This is a magic string. The components of it are
-  # jdbc: the standard java database connection uri scheme
-  # tc: TestContainers - the tool that starts a new container per test
-  # postgresql: which container to start
-  # 12.5: which version of postgres to start
-  # databasename: the name of the db to connect to - any string is okay.
-  default.url = "jdbc:tc:postgresql:12.5:///databasename"
+  default.driver = "org.h2.Driver"
+  default.url = "jdbc:h2:mem:play-test-1279561349;MODE=PostgreSQL"
 }
+play.evolutions.db.default.enabled = "false"
