+++ universal-application-tool-0.0.1/test/models/ProgramTest.java
@@ -8,12 +8,12 @@ import java.util.Locale;
 import java.util.Optional;
 import org.junit.Test;
 import repository.ProgramRepository;
-import repository.WithTruncatingTables;
+import repository.WithPostgresContainer;
 import services.program.BlockDefinition;
 import services.program.ProgramDefinition;
 import services.question.QuestionDefinition;
 
-public class ProgramTest extends WithTruncatingTables {
+public class ProgramTest extends WithPostgresContainer {
 
   @Test
   public void canSaveProgram() {
