+++ universal-application-tool-0.0.1/test/repository/ProgramRepositoryTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import models.Program;
 import org.junit.Test;
 
-public class ProgramRepositoryTest extends WithTruncatingTables {
+public class ProgramRepositoryTest extends WithPostgresContainer {
 
   @Test
   public void createProgram() {
